removed all doubles from game.  This breaks the torpedoes, and sublight navigation.  Both those will be fixed.

+++ StarTrek_KG/Interfaces/IOutputWrite.cs
@@ -39,7 +39,7 @@ namespace StarTrek_KG.Interfaces
         void Prompt(Ship playerShip, string mapText, Game game);
         void Panel(string panelHead, IEnumerable<string> strings);
         string GetPanelHead(string shipName);
-        bool PromptUser(string promptMessage, out double value);
+        bool PromptUser(string promptMessage, out int value);
         bool PromptUser(string promptMessage, out string value);
 
         void OutputConditionAndWarnings(Ship ship, int shieldsDownLevel);
