+++ StarTrek_KG/Subsystem/SubSystem_Base.cs
@@ -18,7 +18,7 @@ namespace StarTrek_KG.Subsystem
         #region Properties
 
             public int Damage { get; set; }
-            public double MaxTransfer { get; set; }
+            public int MaxTransfer { get; set; }
             public SubsystemType Type { get; set; }
 
         #endregion
